Chuck - Licensing modified sources included
* Added original LICENSE to 3party-licenses
* Added reference on our LICENSE
* Updated copyright profile text
* Relocated sources and refactored usages
