+++ .idea/copyright/Chuck_NOTICE.xml
@@ -1,0 +1,6 @@
+<component name="CopyrightManager">
+  <copyright>
+    <option name="notice" value="This source file is part of InAppDevTools, which is available under &#10;Apache License, Version 2.0 at https://github.com/rafaco/InAppDevTools&#10;&#10;This is a modified source from project Chuck, which is available under &#10;Apache License, Version 2.0 at https://github.com/jgilfelt/chuck&#10;&#10;Modifications:&#10;&#10;    * Added this attribution notice&#10;&#10;&#10;Modifications copyright 2018-&amp;#36;{today.year} Rafael Acosta Alvarez&#10;Original copyright (C) 2017 Jeff Gilfelt.&#10;&#10;Licensed under the Apache License, Version 2.0 (the &quot;License&quot;);&#10;you may not use this file except in compliance with the License.&#10;You may obtain a copy of the License at&#10;&#10;    http://www.apache.org/licenses/LICENSE-2.0&#10;&#10;Unless required by applicable law or agreed to in writing, software&#10;distributed under the License is distributed on an &quot;AS IS&quot; BASIS,&#10;WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.&#10;See the License for the specific language governing permissions and&#10;limitations under the License." />
+    <option name="myName" value="Chuck NOTICE" />
+  </copyright>
+</component>
