+++ 3party-licenses/codeview-LICENSE
@@ -1,4 +1,4 @@
-MIT License
+The MIT License (MIT)
 
 Copyright (c) 2016-2017 Tiago Melo
