+++ LICENSE
@@ -212,13 +212,17 @@ This product bundles modified sources of CodeView from Tiago Melo, which is avai
 MIT License. For details, see 3party-licenses/codeview-LICENSE
 Change log per sources at br.tiagohm namespace.
 
-This product bundles modified source of MaterialDesign-Toast from Valdese Kamdem,
+This product bundles modified sources of MaterialDesign-Toast from Valdese Kamdem,
 which is available without license.
 Change log per sources at com.valdesekamdem namespace.
 
-This product bundles modified source of WCViewPager from rnevet, which is available under
+This product bundles modified sources of WCViewPager from rnevet, which is available under
 Apache License Version 2.0. For details, see 3party-licenses/wcviewpager-LICENSE
 Change log per sources at nevet.me namespace.
+
+This product bundles modified sources of Chuck from Jeff Gilfelt, which is available under
+Apache License Version 2.0. For details, see 3party-licenses/chuck-LICENSE
+Change log per sources at com.readystatesoftware.chuck namespace.
 
 
 This product bundles Pandora by Linjiang, which is available under
