+++ library/src/main/java/com/readystatesoftware/chuck/sample/HttpBinService.java
@@ -2,7 +2,19 @@
  * This source file is part of InAppDevTools, which is available under
  * Apache License, Version 2.0 at https://github.com/rafaco/InAppDevTools
  *
- * Copyright 2018-2019 Rafael Acosta Alvarez
+ * This is a modified source from project Chuck, which is available under
+ * Apache License, Version 2.0 at https://github.com/jgilfelt/chuck
+ *
+ * Modifications:
+ *
+ *     * Added this attribution notice
+ *     * Renamed from SampleApiService to HttpBinService
+ *     * Make all methods public
+ *     * Include simulation method
+ *
+ *
+ * Modifications copyright 2018-2019 Rafael Acosta Alvarez
+ * Original copyright (C) 2017 Jeff Gilfelt.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,27 +29,12 @@
  * limitations under the License.
  */
 
-package es.rafaco.inappdevtools.demo.api;
-
-/*
- * Copyright (C) 2017 Jeff Gilfelt.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-//package com.readystatesoftware.chuck.sample;
+package com.readystatesoftware.chuck.sample;
 
 import okhttp3.OkHttpClient;
 import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 import retrofit2.http.Body;
@@ -50,7 +47,7 @@ import retrofit2.http.PUT;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
 
-public class SampleApiService {
+public class HttpBinService {
 
     public static HttpbinApi getInstance(OkHttpClient client) {
         Retrofit retrofit = new Retrofit.Builder()
@@ -117,5 +114,40 @@ public class SampleApiService {
         Call<Void> cache(@Header("If-Modified-Since") String ifModifiedSince);
         @GET("/cache/{seconds}")
         Call<Void> cache(@Path("seconds") int seconds);
+    }
+
+    public static void simulation(OkHttpClient client) {
+        HttpBinService.HttpbinApi api = HttpBinService.getInstance(client);
+        Callback<Void> cb = new Callback<Void>() {
+            @Override public void onResponse(Call call, Response response) {}
+            @Override public void onFailure(Call call, Throwable t) { t.printStackTrace(); }
+        };
+        api.get().enqueue(cb);
+        api.post(new HttpBinService.Data("posted")).enqueue(cb);
+        api.patch(new HttpBinService.Data("patched")).enqueue(cb);
+        api.put(new HttpBinService.Data("put")).enqueue(cb);
+        api.delete().enqueue(cb);
+        api.status(201).enqueue(cb);
+        api.status(401).enqueue(cb);
+        api.status(500).enqueue(cb);
+        api.delay(9).enqueue(cb);
+        api.delay(15).enqueue(cb);
+        api.redirectTo("https://http2.akamai.com").enqueue(cb);
+        api.redirect(3).enqueue(cb);
+        api.redirectRelative(2).enqueue(cb);
+        api.redirectAbsolute(4).enqueue(cb);
+        api.stream(500).enqueue(cb);
+        api.streamBytes(2048).enqueue(cb);
+        api.image("image/png").enqueue(cb);
+        api.gzip().enqueue(cb);
+        api.xml().enqueue(cb);
+        api.utf8().enqueue(cb);
+        api.deflate().enqueue(cb);
+        api.cookieSet("v").enqueue(cb);
+        api.basicAuth("me", "pass").enqueue(cb);
+        api.drip(512, 5, 1, 200).enqueue(cb);
+        api.deny().enqueue(cb);
+        api.cache("Mon").enqueue(cb);
+        api.cache(30).enqueue(cb);
     }
 }
